% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynComm.R
\docType{data}
\name{ALGORITHM}
\alias{ALGORITHM}
\alias{algorithm}
\alias{Algorithm}
\title{List of available algorithms.}
\format{A named list with the names of the available algorithms:
 \describe{
  \item{algorithm}{See available algorithms below.}
 }}
\usage{
ALGORITHM$algorithm
}
\description{
An algorithm mainly defines how nodes and/or communities are processed,
when quality measurements occur and what happens to the communities
depending on the value of the quality obtained.
}
\section{Currently supported algorithms}{

\describe{
  \item{Louvain}{is a greedy optimization method to extract communities 
            from large networks by optimizing the density of edges 
            inside communities to edges outside communities.}
}
}

\examples{
ALGORITHM$LOUVAIN

}
\keyword{datasets}
