% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynComm.R
\docType{methods}
\name{DynComm$communityCount}
\alias{DynComm$communityCount}
\alias{communityCount}
\title{communityCount()}
\usage{
communityCount()
}
\value{
an unsigned integer value with the number of communities
}
\description{
Get the number of communities after the last iteration of the algorithm.
}
\examples{
dc<-DynComm(ALGORITHM$LOUVAIN,QUALITY$MODULARITY,parameters)
dc$communityCount()

}
