% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynComm.R
\docType{methods}
\name{DynComm$nodes}
\alias{DynComm$nodes}
\alias{nodes}
\title{nodes(community)}
\usage{
nodes(community)
}
\arguments{
\item{community}{The name of the intended community}
}
\value{
a list of nodes belonging to the given community
}
\description{
Get all nodes belonging to the given community after the last iteration.
}
\examples{
dc<-DynComm(ALGORITHM$LOUVAIN,QUALITY$MODULARITY,parameters)
dc$nodes(6)

}
