% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynComm.R
\docType{methods}
\name{DynComm$quality}
\alias{DynComm$quality}
\alias{quality}
\title{quality()}
\usage{
quality()
}
\value{
a floating point number
}
\description{
Get the quality measurement of the graph after the last iteration of the 
algorithm.
}
\examples{
dc<-DynComm(ALGORITHM$LOUVAIN,QUALITY$MODULARITY,parameters)
dc$quality()

}
